+++ libs/forms/src/lib/declarative/form-steps/index.ts
@@ -4,5 +4,9 @@ export { FormSteps, type FormStepsProps, type StepPersistenceConfig } from './fo
 export { FormStepsCompletedContent, type FormStepsCompletedContentProps } from './form-steps-completed'
 export { FormStepsContext, type FormStepsContextValue, type StepInfo, useFormStepsContext } from './form-steps-context'
 export { FormStepsIndicator, type FormStepsIndicatorProps } from './form-steps-indicator'
-export { FormStepsNavigation, type FormStepsNavigationProps } from './form-steps-navigation'
+export {
+  FormStepsNavigation,
+  type FormStepsNavigationButtonProps,
+  type FormStepsNavigationProps,
+} from './form-steps-navigation'
 export { FormStepsStep, type FormStepsStepProps } from './form-steps-step'